improve Readme
Structure it like a manpage.

Backpedal screaming data loss warning about delete-old-events. I'm not sure all of it was true. Also, it's called "delete-old-events", so users should expect data will be deleted. It was confusing, too: "data loss" means _unintended_ destruction of data but running this command to find and delete old events is a conscious choice.